exchanges/creds: move tests to corresponding files, add protection and segration for Credentials struct & ptr values
