+++ exchanges/credentials_test.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"testing"
 
+	"github.com/thrasher-corp/gocryptotrader/config"
 	"google.golang.org/grpc/metadata"
 )
 
@@ -367,5 +368,53 @@ func TestGetDefaultCredentials(t *testing.T) {
 	b.SetCredentials("test", "", "", "", "", "")
 	if b.GetDefaultCredentials() == nil {
 		t.Fatal("unexpected return")
+	}
+}
+
+func TestSetAPICredentialDefaults(t *testing.T) {
+	t.Parallel()
+
+	b := Base{
+		Config: &config.Exchange{},
+	}
+	b.API.CredentialsValidator.RequiresKey = true
+	b.API.CredentialsValidator.RequiresSecret = true
+	b.API.CredentialsValidator.RequiresBase64DecodeSecret = true
+	b.API.CredentialsValidator.RequiresClientID = true
+	b.API.CredentialsValidator.RequiresPEM = true
+	b.SetAPICredentialDefaults()
+
+	if !b.Config.API.CredentialsValidator.RequiresKey ||
+		!b.Config.API.CredentialsValidator.RequiresSecret ||
+		!b.Config.API.CredentialsValidator.RequiresBase64DecodeSecret ||
+		!b.Config.API.CredentialsValidator.RequiresClientID ||
+		!b.Config.API.CredentialsValidator.RequiresPEM {
+		t.Error("incorrect values")
+	}
+}
+
+// TestGetAuthenticatedAPISupport logic test
+func TestGetAuthenticatedAPISupport(t *testing.T) {
+	t.Parallel()
+
+	base := Base{
+		API: API{
+			AuthenticatedSupport:          true,
+			AuthenticatedWebsocketSupport: false,
+		},
+	}
+
+	if !base.GetAuthenticatedAPISupport(RestAuthentication) {
+		t.Fatal("Expected RestAuthentication to return true")
+	}
+	if base.GetAuthenticatedAPISupport(WebsocketAuthentication) {
+		t.Fatal("Expected WebsocketAuthentication to return false")
+	}
+	base.API.AuthenticatedWebsocketSupport = true
+	if !base.GetAuthenticatedAPISupport(WebsocketAuthentication) {
+		t.Fatal("Expected WebsocketAuthentication to return true")
+	}
+	if base.GetAuthenticatedAPISupport(2) {
+		t.Fatal("Expected default case of 'false' to be returned")
 	}
 }
